refactor(posts): extract write operations to posts/write.ts
